requirements.txt
openai
discord.py
sqlite3
asyncio
youtube_dl
yt_dlp
nacl
youtube_search